try giving suggestions when unexpected indent happens
when it can parse the beginning of a list item or map item after
a map but with too much indentation, it will panic. on the way "out"
it will register all possible indents it's backtracked over, so that
we can give a list of suggestions for nested things

this may also want to differentiate between indents registered for
maps and for lists, so that it doesn't suggest to indent a list entry
with the same amount as a map key, because that'd still be wrong.